refactor: extract preflight helpers and generateRequestId into useMissions.helpers (#14148)
Move 8 pure functions and their supporting constants from the
useMissions.tsx monolith into useMissions.helpers.ts:

  - generateRequestId (+ requestIdCounter)
  - shouldAllowMissingToolWarning
  - shouldSkipClusterPreflight
  - getMissingTools
  - getMissionContextTools
  - resolveMissionToolRequirements
  - buildMissingToolWarning
  - buildMissionToolUnavailableError

Also moved: OPTIONAL_MISSION_TOOL_PATTERNS, MISSION_CONTEXT_TOOL_KEYS,
MISSING_TOOL_WARNING_HEADING, MISSING_TOOL_WARNING_SUFFIX constants.

useMissions.tsx now imports these from helpers instead of defining them
inline, reducing the monolith by ~85 lines (3698 -> 3614).

All functions are stateless and have zero React dependencies, making
them independently unit-testable.
